requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
piper-tts==1.2.0
numpy==1.24.3